Re-organize inter-frame communication code under shared/messaging
Make it easier for future developers to understand the inter-frame communication
code by grouping it in a directory with a short README and an index module that
defines the API for use by other parts of the client.
