+++ src/annotator/sidebar.js
@@ -3,7 +3,7 @@ import Hammer from 'hammerjs';
 import { addConfigFragment } from '../shared/config-fragment';
 import { sendErrorsTo } from '../shared/frame-error-capture';
 import { ListenerCollection } from '../shared/listener-collection';
-import { PortRPC } from '../shared/port-rpc';
+import { PortRPC } from '../shared/messaging';
 
 import { annotationCounts } from './annotation-counts';
 import BucketBar from './bucket-bar';
