+++ src/shared/messaging/index.js
@@ -1,0 +1,4 @@
+export { PortFinder } from './port-finder';
+export { PortProvider } from './port-provider';
+export { PortRPC } from './port-rpc';
+export { isMessageEqual } from './port-util';
