+++ src/shared/messaging/port-util.js
