+++ src/sidebar/services/test/frame-sync-test.js
@@ -116,10 +116,10 @@ describe('FrameSyncService', () => {
     fakeWindow.parent = new FakeWindow();
 
     $imports.$mock({
-      '../../shared/port-finder': {
+      '../../shared/messaging': {
         PortFinder: sinon.stub().returns(fakePortFinder),
+        PortRPC: FakePortRPC,
       },
-      '../../shared/port-rpc': { PortRPC: FakePortRPC },
     });
 
     frameSync = new Injector()
